% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestAICsearch.R
\name{bestAICsearch}
\alias{bestAICsearch}
\title{netClusterParallel}
\usage{
bestAICsearch(
  binaryMatrix,
  minK = 2,
  maxK = 5,
  chiVec = c(0.001, 0.5, 1, 2, 3),
  startseed = 100,
  nIterations = 50,
  AICrange = 100
)
}
\arguments{
\item{binaryMatrix}{Data to be clustered}

\item{minK}{Min number of clusters}

\item{maxK}{Max number of clusters}

\item{chiVec}{Vector of chi values}

\item{startseed}{Seed}

\item{nIterations}{Number of iterations}

\item{AICrange}{AIC range}
}
\value{
list of AIC scrores
}
\description{
Network-based clustering of multiple seeds using parallel computing
}
