% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BBMMclusterEM.R
\name{get_clusters_bernoulli}
\alias{get_clusters_bernoulli}
\title{get_clusters_bernoulli}
\usage{
get_clusters_bernoulli(
  binaryMatrix,
  chi = 0.5,
  k_clust = 5,
  startseed = 100,
  nIterations = 10,
  verbose = FALSE
)
}
\arguments{
\item{binaryMatrix}{Data to be clustered}

\item{chi}{hyperparameter chi}

\item{k_clust}{Number of clusters}

\item{startseed}{Start seed}

\item{nIterations}{number of iterations}

\item{verbose}{set TRUE to display progress}
}
\value{
a list containing the clusterMemberships
}
\description{
Categorical version of Bernoulli mixture model (binary clustering function BBMMclusterEM)
}
