% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{nice_DAG_plot}
\alias{nice_DAG_plot}
\title{nice_DAG_plot}
\usage{
nice_DAG_plot(my_graph, print_direct = TRUE)
}
\arguments{
\item{my_graph}{DAG}

\item{print_direct}{print DAG if TRUE}
}
\description{
DAG visualization
}
