% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical_scoreagainstdag.R
\name{scoreagainstDAG}
\alias{scoreagainstDAG}
\title{Calculating the score of a sample against a DAG}
\usage{
scoreagainstDAG(
  scorepar,
  incidence,
  datatoscore = NULL,
  marginalise = FALSE,
  bdecatCvec = NULL
)
}
\arguments{
\item{scorepar}{an object of class \code{scoreparameters}; see constructor function \code{\link{scoreparameters}}}

\item{incidence}{a square matrix of dimensions equal to the number of variables with entries in \code{\{0,1\}}, representing the adjacency matrix of the DAG against which the score is calculated}

\item{datatoscore}{(optional) a matrix (vector) containing binary (for BDe score) or continuous (for the BGe score) observations (or just one observation)  to be scored; the number of columns should be equal to the number of variables in the Bayesian network, the number of rows should be equal to the number of observations; by default all data from \code{scorepar} parameter is used}

\item{marginalise}{(optional for continuous data), whether to use the posterior mean for scoring (default) or to marginalise over the posterior distribution (more computationally costly)}

\item{bdecatCvec}{specifying categorical variable vector}
}
\value{
the log of the BDe/BGe score of given observations against a DAG
}
\description{
This function calculates the score of a given sample against a DAG represented by its incidence matrix.
}
