% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_plots.R
\name{density_plot}
\alias{density_plot}
\title{density_plot}
\usage{
density_plot(cluster_results, var_selection = NULL)
}
\arguments{
\item{cluster_results}{Cluster results from function get_clusters}

\item{var_selection}{Selected variables to consider, e.g. c(1:5) for first five only}
}
\value{
A density plot of class recordedplot.
}
\description{
Create 2d visualisation of sample fit to Bayesian network clusters
}
