% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterfns.R
\name{netCluster}
\alias{netCluster}
\title{netCluster}
\usage{
netCluster(
  myData,
  kclust = 3,
  nbg = 0,
  itLim = 20,
  EMseeds = 1,
  binaryClust = TRUE
)
}
\arguments{
\item{myData}{Data to be clustered}

\item{kclust}{Number of clusters}

\item{nbg}{Number of covariates}

\item{itLim}{Maximum number of iterations}

\item{EMseeds}{seeds}

\item{binaryClust}{binary clustering before network-based clustering (TRUE by default)}
}
\value{
a list containing the clusterMemberships and "assignprogress"
}
\description{
Network-based clustering
}
