% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{nice_DAG_plot}
\alias{nice_DAG_plot}
\title{nice_DAG_plot}
\usage{
nice_DAG_plot(
  my_DAG,
  print_direct = TRUE,
  node_size = NULL,
  CPDAG = TRUE,
  node_colours = "#fdae61",
  directed = TRUE
)
}
\arguments{
\item{my_DAG}{DAG}

\item{print_direct}{print DAG if TRUE}

\item{node_size}{node size vector}

\item{CPDAG}{if TRUE, then plot CPDAG instead of DAG}

\item{node_colours}{node colours}

\item{directed}{TRUE if nodes should be directed}
}
\description{
DAG visualization
}
