% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{plot_clusters}
\alias{plot_clusters}
\title{plot_clusters}
\usage{
plot_clusters(cluster_results, data = NULL)
}
\arguments{
\item{cluster_results}{Cluster results}

\item{data}{data}
}
\description{
Plot clusters
}
