% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterfns.R
\name{plot_clusters}
\alias{plot_clusters}
\title{plot_clusters}
\usage{
plot_clusters(learned_clusters)
}
\arguments{
\item{learned_clusters}{Output from get_clusters()}
}
\value{
A plot of DAGs
}
\description{
Plot DAGs of each cluster
}
