% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{plot_clusters}
\alias{plot_clusters}
\title{plot_clusters}
\usage{
plot_clusters(
  cluster_results,
  data = NULL,
  node_colours = "#fdae61",
  binary_cols = TRUE
)
}
\arguments{
\item{cluster_results}{Cluster results}

\item{data}{data}

\item{node_colours}{node colours}

\item{binary_cols}{Input can be a vector of the binary variables; if "TRUE", the binary variables are automatically determined; if NULL, then entropy will be used only}
}
\description{
Plot clusters
}
