% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatebinaryBN.R
\name{sampleData}
\alias{sampleData}
\title{sampleData}
\usage{
sampleData(
  k_clust = 3,
  n_vars = 20,
  n_bg = 3,
  n_samples = NULL,
  bgedges = "different",
  equal_cpt_bg = TRUE
)
}
\arguments{
\item{k_clust}{Number of clusters}

\item{n_vars}{Number of variables}

\item{n_bg}{number of conditioned covariates}

\item{n_samples}{number of samples}

\item{bgedges}{type of background edges}

\item{equal_cpt_bg}{specify if conditional probability table of the background edges is constant across clusters}
}
\value{
sampled binary data
}
\description{
Sample binary data from different Bayes nets
}
