% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatebinaryBN.R
\name{sampleData}
\alias{sampleData}
\title{sampleData}
\usage{
sampleData(kclust = 3, Nvars = 20, sseed = 1, samplesizes = c(300, 400, 500))
}
\arguments{
\item{kclust}{Number of clusters}

\item{Nvars}{Number of variables}

\item{sseed}{Seed}

\item{samplesizes}{Sample sizes}
}
\value{
sampled binary data
}
\description{
Sample binary data from different Bayes nets
}
